+++ macgyver-plugin-ci-jenkins/src/main/java/io/macgyver/plugin/ci/jenkins/PostCompletedBuildResultWebhook.java
@@ -21,7 +21,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
@@ -30,7 +29,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 
 import io.macgyver.core.event.DistributedEvent;
-import io.macgyver.core.event.DistributedEventProviderProxy;
+import io.macgyver.core.event.DistributedEventSystem;
 
 /**
  * This is a webhook for the Post Completed Build Results plugin.
@@ -45,7 +44,7 @@ import io.macgyver.core.event.DistributedEventProviderProxy;
 public class PostCompletedBuildResultWebhook {
 
 	@Autowired
-	DistributedEventProviderProxy devent;
+	DistributedEventSystem eventSystem;
 	
 	Logger logger = LoggerFactory.getLogger(PostCompletedBuildResultWebhook.class);
 	ObjectMapper mapper = new ObjectMapper();
@@ -66,7 +65,7 @@ public class PostCompletedBuildResultWebhook {
 		
 		ObjectNode payload = mapper.createObjectNode().put("url", url);
 		DistributedEvent evt = DistributedEvent.create().topic("ci.jenkins.post-build-completed").payload(payload);
-		devent.publish(evt);
+		eventSystem.publish(evt);
 		// we can call back to <url>/api/json to get actual information about the build
 		return ResponseEntity.ok(mapper.createObjectNode().put("status", "ok"));
 	}